+++ src/AgentFrameworkToolkit.Mistral/MistralAgentOptions.cs
@@ -1,6 +1,7 @@
 using Microsoft.Agents.AI;
 using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Logging;
+using static Microsoft.Agents.AI.ChatClientAgentOptions;
 
 namespace AgentFrameworkToolkit.Mistral;
 
@@ -89,6 +90,19 @@ public class MistralAgentOptions
     /// Enable Logging Middleware for custom Logging
     /// </summary>
     public LoggingMiddleware? LoggingMiddleware { get; set; }
+
+    /// <summary>
+    /// Gets or sets a factory function to create an instance of <see cref="ChatMessageStore"/>
+    /// which will be used to store chat messages for this agent.
+    /// </summary>
+    public Func<ChatMessageStoreFactoryContext, ChatMessageStore>? ChatMessageStoreFactory { get; set; }
+
+    /// <summary>
+    /// Gets or sets a factory function to create an instance of <see cref="AIContextProvider"/>
+    /// which will be used to create a context provider for each new thread, and can then
+    /// provide additional context for each agent run.
+    /// </summary>
+    public Func<AIContextProviderFactoryContext, AIContextProvider>? AIContextProviderFactory { get; set; }
 
     /// <summary>
     /// Apply Middleware to the Agent, if needed