+++ src/AgentFrameworkToolkit.OpenRouter/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using JetBrains.Annotations;
+using AgentFrameworkToolkit.OpenAI;
+using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace AgentFrameworkToolkit.OpenRouter;
@@ -29,5 +30,28 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddOpenRouterAgentFactory(this IServiceCollection services, string apiKey)
     {
         return services.AddSingleton(new OpenRouterAgentFactory(apiKey));
+    }
+
+
+    /// <summary>
+    /// Register an OpenRouterEmbeddingFactory as a Singleton
+    /// </summary>
+    /// <param name="services">The IServiceCollection collection</param>
+    /// <param name="connection">Connection Details</param>
+    /// <returns>The ServiceCollection</returns>
+    public static IServiceCollection AddOpenRouterEmbeddingFactory(this IServiceCollection services, OpenRouterConnection connection)
+    {
+        return services.AddSingleton(new OpenRouterEmbeddingFactory(connection));
+    }
+
+    /// <summary>
+    /// Register an OpenRouterEmbeddingFactory as a Singleton
+    /// </summary>
+    /// <param name="services">The IServiceCollection collection</param>
+    /// <param name="apiKey">The API Key</param>
+    /// <returns>The ServiceCollection</returns>
+    public static IServiceCollection AddOpenRouterEmbeddingFactory(this IServiceCollection services, string apiKey)
+    {
+        return services.AddSingleton(new OpenRouterEmbeddingFactory(apiKey));
     }
 }